refactor(frontend): move endpoints to top-level route
- Move endpoints from /projects/endpoints to /endpoints
- Maintain project-scoped views at /projects/[id]/endpoints
- Update all imports and route references
- Update navigation menu to point to /endpoints
- Improve consistency with other top-level routes (tests, metrics, etc)

BREAKING CHANGE: Endpoints route changed from /projects/endpoints to /endpoints
