+++ apps/frontend/src/app/(protected)/endpoints/[identifier]/layout.tsx
