+++ apps/frontend/src/app/(protected)/endpoints/components/EndpointDetail.tsx
