+++ apps/frontend/src/app/(protected)/endpoints/components/EndpointForm.tsx
@@ -346,7 +346,7 @@ export default function EndpointForm() {
       notifications.show('Endpoint created successfully!', {
         severity: 'success',
       });
-      router.push('/projects/endpoints');
+      router.push('/endpoints');
     } catch (error) {
       setError((error as Error).message);
     }
@@ -366,7 +366,7 @@ export default function EndpointForm() {
           <Box sx={{ display: 'flex', gap: 2 }}>
             <Button
               variant="outlined"
-              onClick={() => router.push('/projects/endpoints')}
+              onClick={() => router.push('/endpoints')}
             >
               Cancel
             </Button>
