+++ apps/frontend/src/app/(protected)/endpoints/layout.tsx
