+++ apps/frontend/src/app/(protected)/endpoints/new/layout.tsx
