+++ apps/frontend/src/app/(protected)/endpoints/page.tsx
